refactor(docs): collapse story-examples into features component
- Purpose: merge the standalone story-examples component into the
  features component to simplify the codebase and reduce indirection,
  since features.component had very little separate logic.
- What changed: removed the story-examples component (TS/HTML/SCSS) and
  inlined its template/logic into apps/docs/src/app/pages/docs-page/features.
- Files removed: apps/docs/src/app/custom-elements/story-playground/
  story-examples.component.ts, .html, .scss.
- Files updated: apps/docs/src/app/pages/docs-page/features/features.component.{ts,html}
  (added story loading/computed/resource logic, pipes and FeatureComponent
  import, and inlined story rendering markup).
- Template changes: replaced <fkt-code-highlight/> with <app-code-highlight/>
  and replaced the <fkt-story-examples/> usage with the inlined markup that
  renders either <fkt-markdown/> or <app-feature/> per story.
- Breaking changes / notes: selector was changed from 'fkt-features' to
  'app-features' and the story-examples custom element was removed — update
  any external references if present.
- Rationale (hint): colapsed the story-examples component into features to
  simplify the implementation and avoid an extra indirection layer.
- Suggested follow-up: search the repo for removed selectors/tags and run
  the docs page to verify rendering and routing behavior.
