--- apps/docs/src/app/custom-elements/story-playground/story-examples.component.html
@@ -1,22 +1,0 @@
-@let indexer = storyIndexer();
-@let resolved = storyResolved.value();
-
-@if (indexer && resolved) {
-    @for (story of stories(); track story.name) {
-        @if ((story.level ?? 2) === 3) {
-            <h3 [id]="story.name | pascalToKebab">{{ story.name | pascalToHumanReadable }}</h3>
-        } @else {
-            <h2 [id]="story.name | pascalToKebab">{{ story.name | pascalToHumanReadable }}</h2>
-        }
-
-        @if (story.type === 'introduction') {
-            <fkt-markdown [data]="story.description ?? ''"/>
-        } @else {
-            <fkt-story-renderer
-                [storyName]="story.name"
-                [indexer]="indexer"
-                [storyResolved]="resolved"
-            />
-        }
-    }
-}
